updating box

+++ problems/.box/clean.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python2.7
+#!/usr/bin/env python3
 # encoding: utf-8
 
 from glob import glob
@@ -8,7 +8,7 @@ import os.path
 import subprocess
 
 def clean_backups(name):
-    u"Delete backup files from problem `name`'s tree."
+    "Delete backup files from problem `name`'s tree."
     for dirpath, _, backup_fns in os.walk('%s/' % name):
         for backup_fn in backup_fns:
             if backup_fn.endswith('~'):
@@ -18,9 +18,12 @@ def clean_build(name):
     subprocess.check_call(['rm', '-rf', '%s/build/tests/' % name])
 
 def box_clean(name):
-    u"Clean up problem tree, deleting all generated files."
+    "Clean up problem tree, deleting all generated files."
     for pdf_fn in glob('%s/documents/*.pdf' % name):
-        delete_file(pdf_fn)
+        # preserve files named submission.pdf
+        head_fn, tail_fn=os.path.split(pdf_fn)
+        if tail_fn!='submission.pdf':
+            delete_file(pdf_fn)
     for log_fn in glob('%s/documents/*.log' % name):
         delete_file(log_fn)
     delete_file('%s/%s.in' % (name, name))
@@ -31,17 +34,23 @@ def box_clean(name):
         delete_file(sol_fn)
     for sol_fn in glob('%s/solutions/*/*.class' % name):
         delete_file(sol_fn)
+    for sol_fn in glob('%s/solutions/*/*.pyc' % name):
+        delete_file(sol_fn)
     for sol_fn in glob('%s/attic/*.exe' % name):
         delete_file(sol_fn)
     for sol_fn in glob('%s/attic/*.jexe' % name):
         delete_file(sol_fn)
     for sol_fn in glob('%s/attic/*.class' % name):
         delete_file(sol_fn)
+    for sol_fn in glob('%s/attic/*.pyc' % name):
+        delete_file(sol_fn)
     for sol_fn in glob('%s/checkers/*.exe' % name):
         delete_file(sol_fn)
     for sol_fn in glob('%s/checkers/*.jexe' % name):
         delete_file(sol_fn)
     for sol_fn in glob('%s/checkers/*.class' % name):
+        delete_file(sol_fn)
+    for sol_fn in glob('%s/checkers/*.pyc' % name):
         delete_file(sol_fn)
     clean_backups(name)
     clean_build(name)
